allow skipping the rm/untar with TEST_NOCLEAN envar

+++ bin/test-setup.sh
@@ -30,9 +30,11 @@ fi
 
 pushd $DIR
 
-pushd $DIR/deployments/launcher/target/
-rm -rf indy
-tar -zxvf indy-launcher-*-complete.tar.gz
+if [ "x${TEST_NOCLEAN}" == "x" ]; then
+  pushd $DIR/deployments/launcher/target/
+  rm -rf indy
+  tar -zxvf indy-launcher-*-complete.tar.gz
+fi
 
 if [ "x${TEST_REPOS}" != "x" ]; then
   echo "Copying repository/group definitions from: ${TEST_REPOS}"
